fix warning with conv to int16 when int8 needed

+++ TreeTycoon/Engine/Inc/Drawable.hpp
@@ -23,7 +23,7 @@ public:
 
 	int8_t getUpdatePriority() const final
 	{
-		return ConvertTo<int16_t>( UpdatePriority::Drawable );
+		return ConvertTo<int8_t>( UpdatePriority::Drawable );
 	}
 
 	void setDrawLayer( int16_t layer_ );
